requirements.txt
qiskit-terra>=0.17.0
scipy>=1.4
numpy>=1.17
docplex>=2.21.207
setuptools>=40.1.0
networkx>=2.2,<2.6
dataclasses; python_version < '3.7'
